requirements.txt
Flask
Flask-SQLAlchemy
Flask-Login
python-dotenv
openai
